NO3 Cutscene: Decomp on PSP (no codeshare), data/BSS on US (#2610)
This was tricky!

There are enough differences between the two versions that I think it
makes sense to keep these non-codeshared. Therefore I think this will be
the state I'll leave it in, at least for now.

Pulled in data on US though, so at least there's that. The .h file is
looking nicely trimmed down too.

+++ src/st/cutscene_actor_name_psp.h
@@ -13,7 +13,7 @@ extern u16 actor_names_it[];
 
 // Creates primitives for the actor name at the head of the dialogue
 void DrawCutsceneActorName(u16 actorIndex, Entity* self, u16* actorNames,
-                           u8* arg3, s32* arg4, s32 nActors) {
+                           u8* arg3, u8* arg4, s32 nActors) {
     Primitive* prim;
     u16 ch;
     u16 i;
